#37 Added basic authentication

+++ backend/src/main/java/ca/mcgill/ecse321/museum/controller/DonationRestController.java
@@ -4,12 +4,19 @@ package ca.mcgill.ecse321.museum.controller;
 import ca.mcgill.ecse321.museum.dto.Request.DonationRequestDto;
 import ca.mcgill.ecse321.museum.dto.Response.DonationResponseDto;
 import ca.mcgill.ecse321.museum.model.Donation;
+import ca.mcgill.ecse321.museum.model.Visitor;
+import ca.mcgill.ecse321.museum.repository.VisitorRepository;
 import ca.mcgill.ecse321.museum.service.DonationService;
 import io.swagger.annotations.Api;
 import java.sql.Date;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -24,27 +31,45 @@ import org.springframework.web.bind.annotation.RestController;
 @Api(tags = "Donation")
 public class DonationRestController {
     @Autowired private DonationService donationService;
+    @Autowired private VisitorRepository visitorRepository;
 
     @PostMapping(value = {"/donations"})
+    @PreAuthorize("hasRole('VISITOR')")
     public ResponseEntity<DonationResponseDto> createDonation(
             @RequestBody DonationRequestDto body) {
         if (body == null) {
             return new ResponseEntity<DonationResponseDto>(HttpStatus.BAD_REQUEST);
         }
-        ;
+        
+        // Check if the authenticated user is the donor
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        // Get the email of the authenticated user
+        String authEmail;
+        if (principal instanceof UserDetails) {
+            authEmail = ((UserDetails)principal).getUsername();
+        } else {
+            authEmail = principal.toString();
+        }
+
+        // Get the id of the authenticated user
+        List<Visitor> visitors = visitorRepository.findByEmail(authEmail);
+
         Donation donation =
-                donationService.createDonation(body.getDescription(), body.getDonorID());
+                donationService.createDonation(body.getDescription(), visitors.get(0).getId());
         return new ResponseEntity<DonationResponseDto>(
                 DonationResponseDto.createDto(donation), HttpStatus.CREATED);
     }
 
     @GetMapping(value = {"/donations/{id}"})
+    @PreAuthorize("hasRole('USER')")
     public ResponseEntity<DonationResponseDto> getDonation(@PathVariable Long id) {
         return new ResponseEntity<DonationResponseDto>(
                 DonationResponseDto.createDto(donationService.getDonation(id)), HttpStatus.OK);
     }
 
     @PutMapping(value = {"/donations/validate/{donationId}/{validatorId}"})
+    @PreAuthorize("hasRole('ADMINISTRATOR')")
     public ResponseEntity<DonationResponseDto> validateDonation(
             @PathVariable Long donationId,
             @PathVariable Long validatorId,
@@ -72,7 +97,38 @@ public class DonationRestController {
     }
 
     @DeleteMapping(value = {"/donations/{donationId}"})
+    @PreAuthorize("hasRole('VISITOR')")
     public ResponseEntity<DonationResponseDto> deleteDonation(@PathVariable Long donationId) {
+        
+        Long donorId;
+        try {
+            // Get the donation owner id
+            Donation donation = donationService.getDonation(donationId);
+            donorId = donation.getDonor().getId();
+
+        } catch (Exception e) {
+            return new ResponseEntity<DonationResponseDto>(HttpStatus.NOT_FOUND);
+        }
+            
+        // Check if the authenticated user is the donor
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        // Get the email of the authenticated user
+        String authEmail;
+        if (principal instanceof UserDetails) {
+            authEmail = ((UserDetails)principal).getUsername();
+        } else {
+            authEmail = principal.toString();
+        }
+
+        // Get the id of the authenticated user
+        List<Visitor> person = visitorRepository.findByEmail(authEmail);
+
+        // Check if the authenticated user is the donor
+        if (donorId != person.get(0).getId()) {
+            return new ResponseEntity<DonationResponseDto>(HttpStatus.UNAUTHORIZED);
+        }
+        
         donationService.deleteDonation(donationId);
         return new ResponseEntity<DonationResponseDto>(HttpStatus.OK);
     }
